热点信息列表添加时覆盖之前的数据

+++ netInsight-alert/src/main/java/com/trs/netInsight/widget/report/service/impl/ReportServiceNewImpl.java
@@ -455,9 +455,11 @@ public class ReportServiceNewImpl implements IReportServiceNew {
 
 				if (isResourceSim) {
 					List<ReportResource> thisChapterList = reportResourceRepository.findByTemplateIdAndChapter(templateId, chapter);
-					if(thisChapterList.size()>0){
-						for(ReportResource reportResource : thisChapterList)
-							reportResourceRepository.delete(reportResource);
+					if("Hot_News_List".equals(chapter)) {
+						if (thisChapterList.size() > 0) {
+							for (ReportResource reportResource : thisChapterList)
+								reportResourceRepository.delete(reportResource);
+						}
 					}
 					List<ReportResource> newChapterList = reportResourceRepository.findByTemplateIdAndChapter(templateId, chapter);
 					Integer docPosition = 0;
